backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pandas==2.2.2
numpy==1.26.4
shapely==2.0.4
httpx==0.27.0
google-generativeai==0.7.2
python-telegram-bot==21.3
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.3.0
scipy==1.13.1
